pointer position calculation properly fixed

+++ Source/Urho3D/Graphics/OpenGL/OGLGraphics.cpp
@@ -72,9 +72,9 @@ extern "C"
 // Helper functions to support emscripten canvas resolution change
 static const Urho3D::Context *appContext;
 
-static void JSCanvasSize(int width, int height, bool fullscreen, int scale)
+static void JSCanvasSize(int width, int height, bool fullscreen, float scale)
 {
-    URHO3D_LOGINFOF("JSCanvasSize: width=%d height=%d fullscreen=%d ui scale=%d", width, height, fullscreen, scale);
+    URHO3D_LOGINFOF("JSCanvasSize: width=%d height=%d fullscreen=%d ui scale=%f", width, height, fullscreen, scale);
 
     using namespace Urho3D;
